+++ lib/src/controls/inputs/checkbox.dart
@@ -1,7 +1,21 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 import 'package:fluent_ui/fluent_ui.dart';
+import '../../utils/focus.dart';
 
+/// A check box is used to select or deselect action items. It can
+/// be used for a single item or for a list of multiple items that
+/// a user can choose from. The control has three selection states:
+/// unselected, selected, and indeterminate. Use the indeterminate
+/// state when a collection of sub-choices have both unselected and
+/// selected states.
+///
+/// ![Checkbox Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/templates-checkbox-states-default.png)
+///
+/// See also:
+/// - [ToggleSwitch](https://pub.dev/packages/fluent_ui#toggle-switches)
+/// - [RadioButton](https://pub.dev/packages/fluent_ui#radio-buttons)
+/// - [ToggleButton]
 class Checkbox extends StatelessWidget {
   const Checkbox({
     Key? key,
@@ -12,7 +26,17 @@ class Checkbox extends StatelessWidget {
     this.focusNode,
   }) : super(key: key);
 
+  /// Whether this [Checkbox] is checked or not.
+  ///
+  /// If this is `null`, the checkbox is in its third state.
   final bool? checked;
+
+  /// Called when the value of the [Checkbox] should change.
+  ///
+  /// This callback passes a new value, but doesn't update its
+  /// state internally.
+  ///
+  /// If this callback is null, the Checkbox is disabled.
   final ValueChanged<bool?>? onChanged;
 
   final CheckboxStyle? style;
@@ -80,14 +104,10 @@ class Checkbox extends StatelessWidget {
             }(),
           ),
         );
-        if (state.isFocused) {
-          child = Container(
-            decoration: BoxDecoration(
-              border: focusedButtonBorder(context.theme),
-            ),
-            child: child,
-          );
-        }
+        child = FocusBorder(
+          focused: state.isFocused,
+          child: child,
+        );
         return Semantics(
           checked: checked,
           child: child,