+++ lib/src/controls/inputs/radio_button.dart
@@ -1,4 +1,5 @@
 import 'package:fluent_ui/fluent_ui.dart';
+import 'package:fluent_ui/src/utils/focus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
@@ -9,14 +10,29 @@ class RadioButton extends StatelessWidget {
     required this.onChanged,
     this.style,
     this.semanticsLabel,
+    this.focusNode,
+    this.autofocus = false,
   }) : super(key: key);
 
+  /// Whether the button is checked of not
   final bool checked;
+
+  /// Called when the value of the button is changed.
+  /// If this is `null`, the button is considered disabled
   final ValueChanged<bool>? onChanged;
 
+  /// The style of the button.
+  /// This is merged with [Style.radioButtonStyle]
   final RadioButtonStyle? style;
 
+  /// The semantics label of the button
   final String? semanticsLabel;
+
+  /// The [FocusNode] of the button
+  final FocusNode? focusNode;
+
+  /// Whether the button should be autofocused
+  final bool autofocus;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -35,6 +51,9 @@ class RadioButton extends StatelessWidget {
     debugCheckHasFluentTheme(context);
     final style = context.theme.radioButtonStyle?.copyWith(this.style);
     return HoverButton(
+      cursor: style?.cursor,
+      autofocus: autofocus,
+      focusNode: focusNode,
       semanticsLabel: semanticsLabel,
       onPressed: onChanged == null ? null : () => onChanged!(!checked),
       builder: (context, state) {
@@ -46,14 +65,10 @@ class RadioButton extends StatelessWidget {
               ? style?.checkedDecoration!(state)
               : style?.uncheckedDecoration!(state),
         );
-        if (state.isFocused) {
-          child = Container(
-            decoration: BoxDecoration(
-              border: focusedButtonBorder(context.theme),
-            ),
-            child: child,
-          );
-        }
+        child = FocusBorder(
+          focused: state.isFocused,
+          child: child,
+        );
         return Semantics(
           child: child,
           selected: checked,