+++ lib/src/controls/inputs/toggle_switch.dart
@@ -2,6 +2,27 @@ import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
+import '../../utils/focus.dart';
+
+/// The toggle switch represents a physical switch that allows users to
+/// turn things on or off, like a light switch. Use toggle switch controls
+/// to present users with two mutually exclusive options (such as on/off),
+/// where choosing an option provides immediate results.
+///
+/// Use a toggle switch for binary operations that take effect right after the
+/// user flips the toggle switch
+///
+/// ![ToggleSwitch Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/toggleswitches01.png)
+///
+/// Think of the toggle switch as a physical power switch for a device: you flip
+/// it on or off when you want to enable or disable the action performed by the device.
+///
+/// See also:
+///
+/// - [Checkbox](https://pub.dev/packages/fluent_ui#checkbox)
+/// - [RadioButton](https://pub.dev/packages/fluent_ui#radio-buttons)
+/// - [ToggleButton](https://pub.dev/packages/fluent_ui#toggle-button)
+/// - [RadioButton](https://github.com/bdlukaa/fluent_ui#radio-buttons)
 class ToggleSwitch extends StatelessWidget {
   const ToggleSwitch({
     Key? key,
@@ -11,17 +32,36 @@ class ToggleSwitch extends StatelessWidget {
     this.semanticsLabel,
     this.thumb,
     this.focusNode,
+    this.autofocus = false,
   }) : super(key: key);
 
+  /// Whether the [ToggleSwitch] is checked
   final bool checked;
+
+  /// Called when the value of the [ToggleSwitch] should change.
+  ///
+  /// This callback passes a new value, but doesn't update its state
+  /// internally.
+  ///
+  /// If this callback is null, the ToggleSwitch is disabled.
   final ValueChanged<bool>? onChanged;
 
+  /// The thumb of this [ToggleSwitch]. If this is null, defaults to [DefaultToggleSwitchThumb]
   final Widget? thumb;
 
+  /// The style of this [ToggleSwitch].
+  ///
+  /// This style is mescled with [Style.toggleSwitchStyle]
   final ToggleSwitchStyle? style;
 
+  /// The `semanticsLabel` of this [ToggleSwitch]
   final String? semanticsLabel;
+
+  /// The focus node of this [ToggleSwitch]
   final FocusNode? focusNode;
+
+  /// Whether this [ToggleSwitch] is auto focused
+  final bool autofocus;
 
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
@@ -36,6 +76,11 @@ class ToggleSwitch extends StatelessWidget {
       onChanged,
       ifNull: 'disabled',
     ));
+    properties.add(FlagProperty(
+      'autofocus',
+      value: autofocus,
+      ifFalse: 'manual focus',
+    ));
     properties.add(DiagnosticsProperty<ToggleSwitchStyle>('style', style));
     properties.add(StringProperty('semanticsLabel', semanticsLabel));
     properties.add(ObjectFlagProperty<FocusNode>.has('focusNode', focusNode));
@@ -44,24 +89,25 @@ class ToggleSwitch extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     debugCheckHasFluentTheme(context);
-    final style = context.theme.toggleSwitchStyle!.copyWith(this.style);
+    final style = context.theme.toggleSwitchStyle?.copyWith(this.style);
     return HoverButton(
+      autofocus: autofocus,
       semanticsLabel: semanticsLabel,
-      margin: style.margin,
+      margin: style?.margin,
       focusNode: focusNode,
-      cursor: style.cursor,
+      cursor: style?.cursor,
       onPressed: onChanged == null ? null : () => onChanged!(!checked),
       builder: (context, state) {
         Widget child = AnimatedContainer(
           alignment: checked ? Alignment.centerRight : Alignment.centerLeft,
           height: 20,
           width: 45,
-          duration: style.animationDuration!,
-          curve: style.animationCurve!,
-          padding: style.padding,
+          duration: style?.animationDuration ?? Duration.zero,
+          curve: style?.animationCurve ?? Curves.linear,
+          padding: style?.padding,
           decoration: checked
-              ? style.checkedDecoration!(state)
-              : style.uncheckedDecoration!(state),
+              ? style?.checkedDecoration?.call(state)
+              : style?.uncheckedDecoration?.call(state),
           child: thumb ??
               DefaultToggleSwitchThumb(
                 checked: checked,
@@ -69,14 +115,10 @@ class ToggleSwitch extends StatelessWidget {
                 state: state,
               ),
         );
-        if (state.isFocused) {
-          child = Container(
-            decoration: BoxDecoration(
-              border: focusedButtonBorder(context.theme),
-            ),
-            child: child,
-          );
-        }
+        child = FocusBorder(
+          child: child,
+          focused: state.isFocused,
+        );
         return Semantics(
           child: child,
           checked: checked,
@@ -95,14 +137,14 @@ class DefaultToggleSwitchThumb extends StatelessWidget {
   }) : super(key: key);
 
   final bool checked;
-  final ToggleSwitchStyle style;
+  final ToggleSwitchStyle? style;
   final ButtonStates state;
 
   @override
   Widget build(BuildContext context) {
     return AnimatedContainer(
-      duration: style.animationDuration ?? Duration.zero,
-      curve: style.animationCurve!,
+      duration: style?.animationDuration ?? Duration.zero,
+      curve: style?.animationCurve ?? Curves.linear,
       constraints: BoxConstraints(
         minHeight: 8,
         minWidth: 8,
@@ -110,8 +152,8 @@ class DefaultToggleSwitchThumb extends StatelessWidget {
         maxWidth: 12,
       ),
       decoration: checked
-          ? style.checkedThumbDecoration!(state)
-          : style.uncheckedThumbDecoration!(state),
+          ? style?.checkedThumbDecoration?.call(state)
+          : style?.uncheckedThumbDecoration?.call(state),
     );
   }
 }
@@ -170,7 +212,7 @@ class ToggleSwitchStyle with Diagnosticable {
       },
       padding: EdgeInsets.symmetric(horizontal: 3, vertical: 4),
       margin: EdgeInsets.all(4),
-      animationDuration: style.mediumAnimationDuration,
+      animationDuration: style.fastAnimationDuration,
       animationCurve: style.animationCurve,
       checkedThumbDecoration: (_) => defaultThumbDecoration.copyWith(color: () {
         if (style.brightness == Brightness.light)